Write block/inode bitmaps and inode tables earlier.
Before we would write blank block/inode bitmap tables then
fill them in later.  This is a bit of a hack to set up enough
of the data structures (I think) to allow the correct block/inode
bitmaps to be written first time.

This may not be worth the hack/risk - I haven't tested the
speed improvement carefully.